backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
Pillow==11.1.0
numpy
opencv-python-headless
imageio
imageio-ffmpeg
sqlalchemy==2.0.36
aiofiles==24.1.0
